feat : 드롭다운 구현, 선택된 상태의 광고만 보여주도록 로직 구현
